requirements.txt
keyring>=24.0.0
Pillow>=10.0.0
numpy>=1.26.0
openai>=1.0.0
